convert to double

+++ Hello world/Program.cs
@@ -12,9 +12,9 @@ int age = 2022 - year;
 Console.WriteLine($"Вам {age} лет!");
 
 Console.WriteLine("Вводите первое чосло?");
-double number = Convert.ToInt32(Console.ReadLine());
+double number = Convert.ToDouble(Console.ReadLine());
 Console.WriteLine("Вводите второе чосло?");
-double number1 = Convert.ToInt32(Console.ReadLine());
+double number1 = Convert.ToDouble(Console.ReadLine());
 double amont = number + number1;
 
 Console.WriteLine($"Суммаы {amont}");
